+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add search field to data tables (gh#aquarist-labs/s3gw#157).
 - Display progress while bootstrapping the app (gh#aquarist-labs/s3gw#247).
+- Make datatable pagination settings persistent (gh#aquarist-labs/s3gw#242).
 
 ## [0.9.0] - 2022-12-01